Les limites pour chaque cases de la grille sont mauvaises et n'arrête pas de changer
